refactor(useEventSource): use `shallowRef` for event & allow any data
Uses `shallowRef` for the event names as they are always strings.

Uses `deepRef` for the event data as it could be anything.

This also raised that the `data` is currently typed as `string | null`,
but event source events can actually pass data of any type. This
compiled successfully in the past because a `MessageEvent` defaults to
`MessageEvent<any>` (i.e. `{data: any}`).